Update nbody.cpp
Fixed the parallelisation of the forces computation using mutexes.
Note: when running ./nbody -comp we still get that the bodies are different when using the sequential and the parallel way. However, the difference is very small (e-05) and is not on all steps (for example in step 100 the bodies are identical). I believe that this is because of accumulated small floating-point rounding inaccuracies. 
When running ./nbody -comp with num_threads = 1, we can see that the bodies are identical in all steps which confirms that the computing of the forces itself is correct.